refactor: merge book processing into job processor
