+++ src/storytime/services/job_processor.py
@@ -1,8 +1,9 @@
-"""Simple job processor for single-voice text-to-audio conversion."""
+"""Unified job processor handling both simple and book jobs."""
 
 import logging
 from datetime import datetime
 from typing import Any
+from uuid import uuid4
 
 from sqlalchemy import select, update
 from sqlalchemy.ext.asyncio import AsyncSession
@@ -10,6 +11,7 @@ from sqlalchemy.ext.asyncio import AsyncSession
 from storytime.database import Job, JobStatus, JobStep, StepStatus
 from storytime.infrastructure.spaces import SpacesClient
 from storytime.models import JobResponse, JobStepResponse
+from storytime.services.book_analyzer import BookAnalyzer, ChapterInfo
 from storytime.services.preprocessing_service import PreprocessingService
 from storytime.services.tts_generator import TTSGenerator
 from storytime.services.web_scraping import WebScrapingService
@@ -18,7 +20,7 @@ logger = logging.getLogger(__name__)
 
 
 class JobProcessor:
-    """Simple processor for text-to-audio jobs."""
+    """Process both simple text jobs and full book jobs."""
 
     def __init__(
         self,
@@ -33,6 +35,16 @@ class JobProcessor:
         self.tts_generator = tts_generator or TTSGenerator()
         self.preprocessing_service = preprocessing_service or PreprocessingService()
         self.web_scraping_service = web_scraping_service or WebScrapingService()
+        self.book_analyzer = BookAnalyzer()
+
+    def _is_book_job(self, job: Job) -> bool:
+        """Determine if a job should be processed as a book."""
+        if job.config:
+            if job.config.get("processing_mode") == "book_splitting":
+                return True
+            if job.config.get("job_type") == "book_processing":
+                return True
+        return False
 
     async def process_job(self, job_id: str) -> JobResponse:
         """Process a job based on its type."""
@@ -47,8 +59,11 @@ class JobProcessor:
         await self._update_job_status(job_id, JobStatus.PROCESSING, started_at=datetime.utcnow())
 
         try:
-            # Process as unified text-to-audio job
-            result = await self._process_text_to_audio_job(job)
+            # Route based on job type
+            if self._is_book_job(job):
+                result = await self._process_book_job(job)
+            else:
+                result = await self._process_text_to_audio_job(job)
 
             # Update job status to completed
             await self._update_job_status(
@@ -68,6 +83,112 @@ class JobProcessor:
                 job_id, JobStatus.FAILED, error_message=str(e), completed_at=datetime.utcnow()
             )
             raise
+
+    async def _process_book_job(self, job: Job) -> dict[str, Any]:
+        """Process a full book job with chapter splitting."""
+        logger.info(f"Processing book job {job.id}")
+
+        # Step 1: Load book text
+        load_step = await self._create_job_step(
+            job.id, "load_book", 0, "Load book text from storage"
+        )
+        await self._update_job_step(load_step.id, StepStatus.RUNNING, started_at=datetime.utcnow())
+
+        try:
+            book_text = await self._load_book_text(job)
+            await self._update_job_step(
+                load_step.id, StepStatus.COMPLETED, progress=1.0, completed_at=datetime.utcnow()
+            )
+        except Exception as e:  # pragma: no cover - basic error propagation
+            await self._update_job_step(
+                load_step.id,
+                StepStatus.FAILED,
+                error_message=str(e),
+                completed_at=datetime.utcnow(),
+            )
+            raise
+
+        # Step 2: Analyze book structure
+        analyze_step = await self._create_job_step(
+            job.id, "analyze_structure", 1, "Analyze book structure and detect chapters"
+        )
+        await self._update_job_step(
+            analyze_step.id, StepStatus.RUNNING, started_at=datetime.utcnow()
+        )
+
+        try:
+            chapters = self.book_analyzer.analyze_book(book_text)
+            await self._update_job_step(
+                analyze_step.id,
+                StepStatus.COMPLETED,
+                progress=1.0,
+                completed_at=datetime.utcnow(),
+                step_metadata={"chapter_count": len(chapters)},
+            )
+        except Exception as e:
+            await self._update_job_step(
+                analyze_step.id,
+                StepStatus.FAILED,
+                error_message=str(e),
+                completed_at=datetime.utcnow(),
+            )
+            raise
+
+        # Step 3: Split and save chapters
+        split_step = await self._create_job_step(
+            job.id, "split_chapters", 2, "Split book into individual chapters"
+        )
+        await self._update_job_step(split_step.id, StepStatus.RUNNING, started_at=datetime.utcnow())
+
+        try:
+            chapter_files = await self._split_and_save_chapters(job.id, book_text, chapters)
+            await self._update_job_step(
+                split_step.id, StepStatus.COMPLETED, progress=1.0, completed_at=datetime.utcnow()
+            )
+        except Exception as e:
+            await self._update_job_step(
+                split_step.id,
+                StepStatus.FAILED,
+                error_message=str(e),
+                completed_at=datetime.utcnow(),
+            )
+            raise
+
+        # Step 4: Create child jobs
+        create_jobs_step = await self._create_job_step(
+            job.id, "create_chapter_jobs", 3, "Create processing jobs for each chapter"
+        )
+        await self._update_job_step(
+            create_jobs_step.id, StepStatus.RUNNING, started_at=datetime.utcnow()
+        )
+
+        try:
+            child_job_ids = await self._create_chapter_jobs(job, chapters, chapter_files)
+            await self._update_job_step(
+                create_jobs_step.id,
+                StepStatus.COMPLETED,
+                progress=1.0,
+                completed_at=datetime.utcnow(),
+                step_metadata={"child_job_ids": child_job_ids},
+            )
+        except Exception as e:
+            await self._update_job_step(
+                create_jobs_step.id,
+                StepStatus.FAILED,
+                error_message=str(e),
+                completed_at=datetime.utcnow(),
+            )
+            raise
+
+        await self._update_job_status(job.id, JobStatus.PROCESSING, progress=1.0)
+
+        return {
+            "processing_type": "book_splitting",
+            "chapter_count": len(chapters),
+            "child_job_ids": child_job_ids,
+            "chapter_files": chapter_files,
+            "total_word_count": len(book_text.split()),
+        }
 
     async def _process_text_to_audio_job(self, job: Job) -> dict[str, Any]:
         """Process a simple text-to-audio conversion job."""
@@ -106,7 +227,7 @@ class JobProcessor:
                         "character_count": scraping_result["character_count"],
                         "estimated_words": scraping_result["estimated_words"],
                         "scraped_title": scraping_result.get("title"),
-                    }
+                    },
                 )
             elif job.input_file_key:
                 # Download from file storage
@@ -124,7 +245,7 @@ class JobProcessor:
                 step_metadata={
                     "content_source": content_source,
                     "character_count": len(text_content),
-                }
+                },
             )
 
             # Update overall job progress (content acquired)
@@ -152,7 +273,7 @@ class JobProcessor:
                 preprocessing_step.id,
                 StepStatus.COMPLETED,
                 progress=1.0,
-                completed_at=datetime.utcnow()
+                completed_at=datetime.utcnow(),
             )
 
             # Update overall job progress (preprocessing complete)
@@ -164,7 +285,9 @@ class JobProcessor:
             )
 
             # Update TTS step to running
-            await self._update_job_step(tts_step.id, StepStatus.RUNNING, started_at=datetime.utcnow())
+            await self._update_job_step(
+                tts_step.id, StepStatus.RUNNING, started_at=datetime.utcnow()
+            )
 
             # Get voice configuration
             voice_config = job.config.get("voice_config", {}) if job.config else {}
@@ -216,8 +339,7 @@ class JobProcessor:
                 for step_name in ["text_to_audio", "preprocess_text", "acquire_content"]:
                     step_result = await self.db_session.execute(
                         select(JobStep).where(
-                            JobStep.job_id == job.id,
-                            JobStep.step_name == step_name
+                            JobStep.job_id == job.id, JobStep.step_name == step_name
                         )
                     )
                     step = step_result.scalar_one_or_none()
@@ -225,7 +347,10 @@ class JobProcessor:
                     if step and step.status == StepStatus.RUNNING:
                         # Found the running step that failed
                         await self._update_job_step(
-                            step.id, StepStatus.FAILED, error_message=str(e), completed_at=datetime.utcnow()
+                            step.id,
+                            StepStatus.FAILED,
+                            error_message=str(e),
+                            completed_at=datetime.utcnow(),
                         )
                         break
             except Exception as db_error:
@@ -386,3 +511,125 @@ class JobProcessor:
             duration=job.duration,
             steps=step_responses,
         )
+
+    async def _load_book_text(self, job: Job) -> str:
+        """Load and preprocess book text from various sources."""
+        if job.config and job.config.get("content"):
+            raw_text = job.config["content"]
+        elif job.config and job.config.get("url"):
+            url = job.config["url"]
+            logger.info(f"Scraping content from URL for job {job.id}")
+            scraping_result = await self.web_scraping_service.extract_content(url)
+            raw_text = scraping_result["content"]
+        elif job.input_file_key:
+            raw_text = await self.spaces_client.download_text_file(job.input_file_key)
+        else:
+            raise ValueError("No book content, file, or URL provided")
+
+        logger.info(f"Preprocessing book text for job {job.id}")
+        processed_text = await self.preprocessing_service.preprocess_text(raw_text, job.config)
+        return processed_text
+
+    async def _split_and_save_chapters(
+        self, job_id: str, book_text: str, chapters: list[ChapterInfo]
+    ) -> list[dict[str, Any]]:
+        chapter_files = []
+        for i, chapter in enumerate(chapters):
+            chapter_text = book_text[chapter.start_position : chapter.end_position]
+            chapter_number = chapter.chapter_number or (i + 1)
+            file_key = f"jobs/{job_id}/chapters/chapter_{chapter_number:03d}.txt"
+            await self.spaces_client.upload_text_file(file_key, chapter_text)
+            chapter_files.append(
+                {
+                    "chapter_number": chapter_number,
+                    "title": chapter.title,
+                    "file_key": file_key,
+                    "word_count": chapter.word_count,
+                    "is_special": chapter.is_special,
+                }
+            )
+        return chapter_files
+
+    async def _create_chapter_jobs(
+        self, parent_job: Job, chapters: list[ChapterInfo], chapter_files: list[dict[str, Any]]
+    ) -> list[str]:
+        child_job_ids = []
+        voice_config = parent_job.config.get("voice_config", {}) if parent_job.config else {}
+        processing_mode = (
+            parent_job.config.get("processing_mode", "single_voice")
+            if parent_job.config
+            else "single_voice"
+        )
+
+        for chapter_file in chapter_files:
+            child_job = Job(
+                id=str(uuid4()),
+                user_id=parent_job.user_id,
+                parent_id=parent_job.id,
+                title=f"{parent_job.title} - {chapter_file['title']}",
+                description=f"Chapter {chapter_file['chapter_number']} of {parent_job.title}",
+                status=JobStatus.PENDING,
+                config={
+                    "parent_job_id": parent_job.id,
+                    "chapter_number": chapter_file["chapter_number"],
+                    "voice_config": voice_config,
+                    "processing_mode": processing_mode,
+                },
+                input_file_key=chapter_file["file_key"],
+            )
+            self.db_session.add(child_job)
+            child_job_ids.append(child_job.id)
+
+        await self.db_session.commit()
+
+        from storytime.worker.tasks import process_job
+
+        for child_id in child_job_ids:
+            try:
+                process_job.delay(child_id)
+            except Exception as e:  # pragma: no cover - scheduling may fail in tests
+                logger.warning(f"Could not schedule child job {child_id}: {e}")
+
+        return child_job_ids
+
+    async def aggregate_chapter_results(self, job_id: str) -> dict[str, Any]:
+        parent_job = await self._get_job(job_id)
+        if not parent_job:
+            raise ValueError(f"Parent job {job_id} not found")
+
+        child_job_ids = (
+            parent_job.result_data.get("child_job_ids", []) if parent_job.result_data else []
+        )
+        if not child_job_ids:
+            logger.warning(f"No child jobs found for parent job {job_id}")
+            return {}
+
+        result = await self.db_session.execute(select(Job).where(Job.id.in_(child_job_ids)))
+        child_jobs = result.scalars().all()
+
+        completed = []
+        failed = []
+        total_duration = 0.0
+        for child in sorted(child_jobs, key=lambda j: j.config.get("chapter_number", 0)):
+            info = {
+                "chapter_number": child.config.get("chapter_number"),
+                "title": child.title,
+                "status": child.status,
+            }
+            if child.status == JobStatus.COMPLETED:
+                info["audio_file"] = child.output_file_key
+                info["duration"] = child.duration
+                if child.duration:
+                    total_duration += child.duration
+                completed.append(info)
+            elif child.status == JobStatus.FAILED:
+                info["error"] = child.error_message
+                failed.append(info)
+
+        return {
+            "total_chapters": len(child_jobs),
+            "completed_chapters": len(completed),
+            "failed_chapters": len(failed),
+            "total_duration_seconds": total_duration,
+            "chapters": completed + failed,
+        }
